Move the hooks to new files.
